+++ client/row_collection/collections.js
@@ -1,3 +1,13 @@
+var afterClearCollection = function (error, result, collectionName, sessionKey) {
+  if (!error) {
+	ConstellationDict.set(sessionKey, 0);
+	console.log(result); 
+  }
+  else {
+	Constellation.error('clearCollection');  
+  }
+}
+
 Template.Constellation_collection_count.helpers({
   collectionCount: function () {
 
@@ -21,15 +31,19 @@ Template.Constellation_collection_count.helpers({
 
   },
   autopublished: function () {
+	var collectionName = String(this);
 	if (ConstellationDict.get('Constellation_autopublish_all')) {
+	  if ( !Package['constellation:autopublish'] || Constellation.collectionIsLocal(collectionName)) {
+        return '';
+	  }
       var notAutopublished = ConstellationDict.get('Constellation_not_autopublished') || [];
-	  return (!_.contains(notAutopublished, String(this))) ? 'Constellation_autopublished' : 'Constellation_not_autopublished';	
+	  return (!_.contains(notAutopublished, collectionName)) ? 'Constellation_autopublished' : 'Constellation_not_autopublished';	
 	}
 	var autopublished = Package['constellation:autopublish'] && ConstellationDict.get('Constellation_autopublished');
-	return (_.contains(autopublished, String(this))) ? 'Constellation_autopublished' : '';
+	return (_.contains(autopublished, collectionName) && collectionName && !Constellation.collectionIsLocal(collectionName)) ? 'Constellation_autopublished' : '';
   },
   toggleAutopublish: function () {
-	return Package["constellation:autopublish"];  
+	return Package["constellation:autopublish"] && !Constellation.collectionIsLocal(String(this));  
   }
 });
 
@@ -48,6 +62,31 @@ Template.Constellation_collection_count.events({
 	}
 	else {
 	  ConstellationDict.set('Constellation_autopublished', autopublishedOrNot);
+	}
+  },
+  'click .Constellation_clear_collection' : function (evt, tmpl) {
+	evt.stopPropagation();
+	var CollectionName = String(tmpl.data);
+	var sessionKey = Constellation.sessKey(CollectionName);
+	if (confirm("This will clear the whole '" + CollectionName + "' collection.\n\nThis cannot be undone.\n\nAre you sure?")) {
+	  if (Constellation.collectionIsLocal(CollectionName)) {
+		// Just make a duplicate on the client
+		var error = null;
+		var result = null;
+		try {
+		  result = Constellation.clearCollection(CollectionName);
+		}
+		catch (err) {
+		  error = err;  
+		}
+		if (!error) {
+		  afterClearCollection.call(null, error, result, CollectionName, sessionKey);
+		}
+		return;    
+	  }
+	  Meteor.call('Constellation_clear_collection', CollectionName, function (error, result) {
+		afterClearCollection.call(null, error, result, CollectionName, sessionKey); 
+	  });	
 	}
   }
 });