+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name:    'constellation:console',
   summary: 'An extensible development console for Meteor',
-  version: '1.2.3',
+  version: '1.3.0',
   git:     'https://github.com/JackAdams/constellation.git',
   documentation: 'README.md',
   debugOnly: true
@@ -50,8 +50,8 @@ Package.onUse(function(api) {
     "common/mutators.js"
   ];
 
-  api.versionsFrom('1.1');
-  api.use(['templating','tracker','mongo','session','underscore','reactive-var','reactive-dict','blaze'], 'client');
+  api.versionsFrom('1.0');
+  api.use(['templating','tracker','mongo','session','underscore','reactive-var','reactive-dict','blaze','jquery'], 'client');
   api.use('check');
   api.use('aldeed:collection2@2.5.0', {weak: true}); // This must go before: api.use('dburles:mongo-collection-instances@0.3.4');
   api.use('dburles:mongo-collection-instances@0.3.4');